(bug) reset token renewal to one hour
In version v0.45.0 token renewal was set to one hour. In release
v0.46.0 this was increased to three hours.
Since the new value was larger than older one, after upgrade Sveltos
did not renew the token for three hours. But the previously generated
token (created during v0.45.0) was set to one hour. So by the time,
after upgrade, Sveltos tried to renew the token it was already expired.