(bug) keep Secret and SveltosCluster in sync
Each SveltosCluster is associated with a Secret that
holds the kubeconfig for accessing the cluster.

The SveltosCluster object includes a field, __Spec.KubeconfigKeyName__,
which specifies the key within the Secret where the kubeconfig is stored.

Should a SveltosCluster pod be configured to renew its token, it will:

- Add a new entry to the Secret's Data field.
- Update SveltosCluster.Spec.KubeconfigKeyName to reflect the new key.

During an upgrade, the register-mgmt-cluster process generates a fresh
kubeconfig. It then updates both the Secret and
SveltosCluster.Spec.KubeconfigKeyName to ensure these fields remain
synchronized.